+++ ledger/app_settings.py
@@ -31,11 +31,16 @@ FUZZ_BASE_URL_REGEX = r"^http[s]?:\/\/(www\.)?fuzzwork\.co\.uk\/"
 # Max Time to set Char Inactive
 LEDGER_CHAR_MAX_INACTIVE_DAYS = clean_setting("LEDGER_CHAR_MAX_INACTIVE_DAYS", 3)
 
-# Set the Stale Status for Ledger Updates in Minutes
-LEDGER_STALE_STATUS = clean_setting("LEDGER_STALE_STATUS", 60)
-
 # Set the Corporation Tax for Corporation & CharacterLedger Caluclation of ESS Payout
 LEDGER_CORP_TAX = clean_setting("LEDGER_CORP_TAX", 15)
 
 # Global timeout for tasks in seconds to reduce task accumulation during outages.
 LEDGER_TASKS_TIME_LIMIT = clean_setting("LEDGER_TASKS_TIME_LIMIT", 7200)
+
+LEDGER_STALE_TYPES = {
+    "journal": 60,
+    "balance": 60,
+    "mining": 60,
+    "planets": 60,
+    "planets_details": 60,
+}